# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ForecastPoint": {
        "properties": {
          "compliant": {
            "title": "Compliant",
            "type": "boolean"
          },
          "predicted_chlorine_mgL": {
            "title": "Predicted Chlorine Mgl",
            "type": "number"
          },
          "predicted_cu_ppb": {
            "title": "Predicted Cu Ppb",
            "type": "number"
          },
          "predicted_turbidity_node4": {
            "title": "Predicted Turbidity Node4",
            "type": "number"
          },
          "predicted_zn_ppb": {
            "title": "Predicted Zn Ppb",
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "predicted_cu_ppb",
          "predicted_zn_ppb",
          "predicted_chlorine_mgL",
          "predicted_turbidity_node4",
          "compliant"
        ],
        "title": "ForecastPoint",
        "type": "object"
      },
      "ForecastResponse": {
        "properties": {
          "horizon_hours": {
            "maximum": 24.0,
            "minimum": 1.0,
            "title": "Horizon Hours",
            "type": "integer"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ForecastPoint"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "horizon_hours"
        ],
        "title": "ForecastResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryPoint": {
        "properties": {
          "chlorine_mgL": {
            "title": "Chlorine Mgl",
            "type": "number"
          },
          "coc": {
            "title": "Coc",
            "type": "number"
          },
          "compliant": {
            "title": "Compliant",
            "type": "boolean"
          },
          "cu_ppb": {
            "title": "Cu Ppb",
            "type": "number"
          },
          "ph_node4": {
            "title": "Ph Node4",
            "type": "number"
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recommendation"
              },
              {
                "type": "null"
              }
            ]
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "turbidity_node4": {
            "title": "Turbidity Node4",
            "type": "number"
          },
          "zn_ppb": {
            "title": "Zn Ppb",
            "type": "number"
          }
        },
        "required": [
          "timestamp",
          "coc",
          "cu_ppb",
          "zn_ppb",
          "chlorine_mgL",
          "turbidity_node4",
          "ph_node4",
          "compliant"
        ],
        "title": "HistoryPoint",
        "type": "object"
      },
...
      "Recommendation": {
        "properties": {
          "reasoning": {
            "items": {
              "type": "string"
            },
            "title": "Reasoning",
            "type": "array"
          },
          "recommended_biocide_dose_mgL": {
            "title": "Recommended Biocide Dose Mgl",
            "type": "number"
          },
          "recommended_blowdown_rate_m3_day": {
            "title": "Recommended Blowdown Rate M3 Day",
            "type": "number"
          },
          "recommended_coc_target": {
            "title": "Recommended Coc Target",
            "type": "number"
          },
          "recommended_ph_setpoint": {
            "title": "Recommended Ph Setpoint",
            "type": "number"
          }
        },
        "required": [
          "recommended_blowdown_rate_m3_day",
          "recommended_biocide_dose_mgL",
          "recommended_ph_setpoint",
          "recommended_coc_target"
        ],
        "title": "Recommendation",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "CoolSense API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/forecast": {
      "get": {
        "operationId": "get_forecast_v1_forecast_get",
        "parameters": [
          {
            "in": "query",
            "name": "horizon_hours",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 24,
              "minimum": 1,
              "title": "Horizon Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Forecast",
        "tags": [
          "forecast"
        ]
      }
    },
...
    "/v1/state/forecast": {
      "get": {
        "operationId": "get_forecast_v1_state_forecast_get",
        "parameters": [
          {
            "in": "query",
            "name": "horizon_hours",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 24,
              "minimum": 1,
              "title": "Horizon Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Forecast",
        "tags": [
          "state"
        ]
      }
    },
    "/v1/state/history": {
      "get": {
        "operationId": "get_history_v1_state_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "since_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since Ts"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoryPoint"
                  },
                  "title": "Response Get History V1 State History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get History",
        "tags": [
          "state"
        ]
      }
    }
  }
}